Listas tablas de products con tabla intermedia config

+++ migrations/20240502094158-create-product.js
@@ -19,7 +19,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.FLOAT
+        type: Sequelize.DECIMAL(10,2)
       },
       stock: {
         type: Sequelize.INTEGER
